dom/base/test/test_document_register.html

<!DOCTYPE html>
<html>
  <head>
    <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css">
    <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  </head>
  <body onload="startTests()">
    <a href="https://bugzilla.mozilla.org/show_bug.cgi?id=783129" target="_blank">Mozilla Bug 783129</a>
    <iframe id="fooframe" src="/"></iframe>
    <script type="application/javascript">

    /** Test for Bug 783129 **/
    SimpleTest.waitForExplicitFinish();

    function startTests() {
      var c = document.getElementById("fooframe").contentDocument.registerElement("x-foo");
      var elem = new c();
      is(elem.tagName, "X-FOO", "Constructor should create an x-foo element.");

      var anotherElem = $("fooframe").contentDocument.createElement("x-foo");
      is(anotherElem.tagName, "X-FOO", "createElement should create an x-foo element.");
      SimpleTest.finish();
    }

    </script>
  </body>
</html>
